+++ src/collective/relationhelpers/__init__.py
@@ -2,5 +2,4 @@
 """Init and utils."""
 from zope.i18nmessageid import MessageFactory
 
-
 _ = MessageFactory('collective.relationhelpers')