+++ CHANGES.rst
@@ -10,6 +10,8 @@ Changelog
   * check view-permissions by default
   * add convenience-methods relations, backrelations, unrestricted_relations and unrestricted_backrelations
   * add convenience-method relation for relationChoice that only returns one object, not a list
+  * rename parameter backref to backrel
+  * allow to query for multiple reations
 
 
 1.0a1 (2020-05-29)